# src/lib/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduWeb API",
        "description": "Documentación de la API para la plataforma educativa EduWeb, diseñada para ofrecer una experiencia de aprendizaje en línea integral.",
        "version": "1.0.0"
    },
...
    "paths": {
        "/users": {
            "get": {
                "tags": ["Usuarios"],
                "summary": "Obtener todos los usuarios",
                "description": "Recuperar una lista paginada de usuarios",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuarios obtenidos exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener los usuarios"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Usuarios"],
                "summary": "Crear un nuevo usuario",
                "description": "Registrar un nuevo usuario en el sistema",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/UserInput"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario creado exitosamente"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Entrada inválida o correo/teléfono duplicado"
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear el usuario"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_name": {
                        "type": "string"
                    },
                    "user_lastname": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "user_image": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                }
            },
            "UserInput": {
                "type": "object",
                "properties": {
                    "user_name": {
                        "type": "string"
                    },
                    "user_lastname": {
                        "type": "string"
                    },
                    "user_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "user_password": {
                        "type": "string"
                    },
                    "user_phone": {
                        "type": "string"
                    },
                    "role_id": {
                        "type": "integer"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "role_id": {
                        "type": "integer"
                    },
                    "role_name": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}